Rename rule.go file to builder + move stdoutResults
